requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
about-time=4.2.1=pypi_0
alive-progress=3.1.5=pypi_0
autograd=1.6.2=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py39h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=3.0.0=pypi_0
cma=3.2.2=pypi_0
contourpy=1.2.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
cython=3.0.10=pypi_0
deap=1.4.1=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.8=pypi_0
et-xmlfile=1.1.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=pypi_0
future=1.0.0=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
gpytorch=1.11=pypi_0
grapheme=0.6.0=pypi_0
idna=3.4=py39h06a4308_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jaxtyping=0.2.28=pypi_0
jinja2=3.1.3=py39h06a4308_0
joblib=1.4.0=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=13.2.0=h7e041cc_5
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
linear-operator=0.5.2=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.8.4=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.1.5=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.2.0=py39h5eee18b_0
pip=23.3.1=py39h06a4308_0
ply=3.11=py39h06a4308_0
pymoo=0.6.1.1=pypi_0
pyomo=6.7.1=py39h3d6467e_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=2_cp39
pytorch=2.2.2=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py39h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_1
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.2.2=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.4.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.2.2=py39_cu118
torchtriton=2.2.0=py39
torchvision=0.17.2=py39_cu118
typeguard=2.13.3=pypi_0
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2024.1=pypi_0
urllib3=2.1.0=py39h06a4308_1
wheel=0.41.2=py39h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
